Optional JargonResponseBuilder behavior to combine multiple calls to speak and reprompt
By default the speak and reprompt methods replace the content from previous calls to those methods; this behavior mirrors
that of corresponding ASK SDK methods. There are two ways to change this behavior such to multiple calls to result in content
getting merged (with a space in between) instead of replaced:

1. When constructing the JargonSkillBuilder pass in an options objeect with mergeSpeakAndReprompt set to true
2. Set the merge parameter to the speak or reprompt method to true

When mergeSpeakAndReprompt is true the default replace behavior can be used for specific calls to speak or reprompt by
setting the merge parameter to false.

Note that each indivdual call to `speak` or `reprompt` should contain content that can stand alone (e.g., a full sentence or
paragraph) to minimize the chances that the order of the content would change across langauges.