Initial Voxa plugin implementation
The 'JargonVoxa' function handles creating the VoxaApp, configured with the
Jargon renderer and plugin (request handler). The Jargon request handler takes
care of creating a ResourceManager for the request's locale, and places it onto
the IVoxaEvent for the request. JargonRenderer uses the resource manager to render
the resources for the response. If desired options for that render operation can
be placed onto the event.

Includes tests for basic functionality. Those tests do not yet cover all of the
different directive and transition types, in particular transitions that specify
a reply object, or a view containing multiple directives.

All rendering is using ResourceManager.renderObject, so variations aren't yet
supported; that will happen for a 1.0 release.