Support returing render items from voxa variables
This builds on the just added support for nested render items in
ResourceManager.renderObject. With this change all variable functions
are always called during rendering, even if the variable won't be used
by the specified resource(s). While this is different behavior than
the standard Voxa renderer, it's also something we'd need to do when
performing remote rendering.

Variable function calls aren't memoized, but it's not clear if we'd gain
much by doing so as those calls should be fast and idempotent.